requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
asyncio==3.4.3
cchardet==2.1.7
kafka-python==2.0.2
psycopg2-binary==2.8.6
pytest==6.2.2
pytest-asyncio==0.14.0
